\name{Test1}
\alias{Test1}
\docType{data}
\title{
BACTERIAL MOVING DISTANCE DATA 
}
\description{
  This dataset gives the moving distance of 18  observations. 3 replications 
  in  each different setting. 
}
\usage{data("Test1")}
\format{
  A data frame with 18 observations on the following 3 variables.
  \describe{
    \item{\code{Strain}}{a character vector containg the name of tested strains}
    \item{\code{Time}}{a numeric vector containing measuring timepoint}
    \item{\code{Distance}}{a numeric vector containing measuring results}
  }
}
\keyword{datasets}