\name{Test2}
\alias{Test2}
\docType{data}
\title{
Bacterial Growth
}
\description{
  This dataset gives the bacterial growth data over 24 h, including 72 observation.
  3 replications in each time point. 
}
\usage{data("Test2")}
\format{
  A data frame with 72 observations on the following 3 variables.
  \describe{
    \item{\code{Strain}}{a character vector}
    \item{\code{Time}}{a numeric vector}
    \item{\code{OD600}}{a numeric vector}
  }
}
\keyword{datasets}
