\name{Test1}
\alias{Test1}
\docType{data}
\title{
Distance between Bacteria and Petri Dish Edge 
}
\description{
  This dataset gives the distance between bacteria and the petri dish of 18 
  observations. 3 replications in each different setting. 
}
\usage{data("Test1")}
\format{
  A data frame with 18 observations on the following 3 variables.
  \describe{
    \item{\code{Strain}}{a character vector}
    \item{\code{Time}}{a numeric vector}
    \item{\code{Distance}}{a numeric vector}
  }
}
\keyword{datasets}